+++ server/login/enter.go
@@ -17,20 +17,7 @@ func init() {
 	Enter = enterLogin{}
 }
 
-func ClientConnStart(conn liFace.IConnection) {
-	app.MClientData.Inc()
-	utils.Log.Info("ClientConnStart:%s", conn.RemoteAddr().String())
-}
 
-func ClientConnStop(conn liFace.IConnection) {
-	app.MClientData.Dec()
-	SessLoginMgr.SessionExitByConn(conn)
-	utils.Log.Info("ClientConnStop:%s", conn.RemoteAddr().String())
-}
-
-func ShutDown(){
-	utils.Log.Info("ShutDown")
-}
 
 type enterLogin struct {
 	liNet.BaseRouter
