+++ server/world/enter.go
@@ -10,21 +10,6 @@ import (
 	"github.com/llr104/LiFrame/utils"
 )
 
-func ClientConnStart(conn liFace.IConnection) {
-	app.MClientData.Inc()
-	utils.Log.Info("ClientConnStart:%s", conn.RemoteAddr().String())
-}
-
-func ClientConnStop(conn liFace.IConnection) {
-	app.MClientData.Dec()
-	app.SessionMgr.SessionExitByConn(conn)
-
-	utils.Log.Info("ClientConnStop:%s", conn.RemoteAddr().String())
-}
-
-func ShutDown(){
-	utils.Log.Info("ShutDown")
-}
 
 var Enter enterWorld
